Mobile Single booking & global availability/unbooking fixes

+++ src/components/mobile/mobileTimeslotConfirmation.tsx
@@ -81,6 +81,7 @@ interface MobileTimeSlotConfirmationProps {
   enabled: boolean;
   date: Date;
   tId: string;
+  allBookings: Booking[];
   setRequery: (requery: boolean) => void;
 }
 
@@ -91,6 +92,7 @@ export default function MobileTimeSlotConfirmation({
   enabled,
   date,
   tId,
+  allBookings,
   setRequery,
 }: MobileTimeSlotConfirmationProps) {
   const currentUserFR = useContext(UserContext);
@@ -188,7 +190,9 @@ export default function MobileTimeSlotConfirmation({
         ])
       );
       bookings.forEach((booking) => {
-        DataStore.delete(booking);
+        if (booking.userID === id) {
+          DataStore.delete(booking);
+        }
       });
     } catch (error: unknown) {
       if (error instanceof Error) {
@@ -215,6 +219,15 @@ export default function MobileTimeSlotConfirmation({
       deleteRVBooking(tId);
       console.log("booked");
     } else {
+      if (
+        userType === "Rider" &&
+        allBookings.some(
+          (booking) =>
+            booking.date === convertToYMD(date) && booking.userID === id
+        )
+      ) {
+        return;
+      }
       addRVBooking(tId, id, date);
       console.log("unbooked");
     }
